Fix crash when reordering items in playlist.
It was caused by adding a new column to the model. When drag started,
generated drag event data contained all ten columns, but recieving
method expected only nine columns. The mouseMoveEvent() which generates
the drag-data is now more robust and will always insert only the columns
it is supposed to insert.