Simplify PlaylistModel
The PlaylistModel is not much more simple and smaller, so the overall
memory footprint of TepSonic should now be awesomely small (and there
are still many places we can improve it).

At this point the model is somehow broken (in combination with the
overlaying proxy model), probably due to issue with mapping from/to
the proxy model.