Make Playlist::loadMetaDataRunnable less blocking
As the entire point of the runnable is to load metadata asynchronously,
it makes no sense to hold a mutex the entire time the runnable runs. We
now instead lock only to modify the metadata and then emit dataChanged().

Since the Playlist model emits dataChanged() for only one column, we need
to map the signal in PlaylistProxyModel to apply to all columns.